% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisco.R
\name{pisco}
\alias{pisco}
\title{values extraction of PISCO data}
\usage{
pisco(x, type = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO file name, longitude and latitude of station.}

\item{type}{extraction type, daily is default value, other option is monthly}
}
\description{
extraction of daily and monthly values from PISCO data, PISCO is Peruvian Interpolated Data of the SENAMHI Climatological and Hydrological Observations.
}
\author{
Geomar Perales Apaico
}
