% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscogroup.R
\name{piscogroup}
\alias{piscogroup}
\title{values extraction of PISCO daily data from a stations group.}
\usage{
piscogroup(x, type = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO file name, longitude and latitude from stations group.}

\item{type}{extraction type, daily is default value, other option is monthly}
}
\description{
extraction of daily and monthly values from PISCO data, PISCO is Peruvian Interpolated Data of the SENAMHI Climatological and Hydrological Observations.
}
\author{
Geomar Perales Apaico
}
